init command line tools

+++ shallow-remote/src/test/java/org/shallow/remote/RemoteClientProcessorAwareTest.java
@@ -14,7 +14,7 @@ public class RemoteClientProcessorAwareTest implements ProcessorAware {
     }
 
     @Override
-    public void process(Channel channel, int command, ByteBuf data, InvokeRejoin<ByteBuf> rejoin) {
+    public void process(Channel channel, byte command, ByteBuf data, InvokeRejoin<ByteBuf> rejoin) {
 
     }
 }
